+++ robotidy/transformers/AddMissingEnd.py
@@ -6,9 +6,10 @@ except ImportError:
     InlineIfHeader = None
 
 from robotidy.disablers import skip_if_disabled, skip_section_if_disabled
+from robotidy.transformers import Transformer
 
 
-class AddMissingEnd(ModelTransformer):
+class AddMissingEnd(Transformer):
     """
     Add missing END token to FOR loops and IF statements.
 
@@ -33,6 +34,9 @@ class AddMissingEnd(ModelTransformer):
 
     Supports global formatting params: ``--startline`` and ``--endline``.
     """
+
+    def __init__(self):
+        super().__init__()  # workaround for our dynamically imported classes with args from cli/config
 
     def fix_block(self, node, expected_type):
         self.generic_visit(node)